Move prepare_pixels to the canvas thread
Nothing else uses it in the whole crate graph.
