+++ components/canvas/Cargo.toml
@@ -15,12 +15,14 @@ webgl_backtrace = ["canvas_traits/webgl_backtrace"]
 
 [dependencies]
 azure = {git = "https://github.com/servo/rust-azure"}
+byteorder = "1"
 canvas_traits = {path = "../canvas_traits"}
 compositing = {path = "../compositing"}
 cssparser = "0.25"
 euclid = "0.19"
 fnv = "1.0"
 gleam = "0.6.4"
+half = "1"
 ipc-channel = "0.11"
 log = "0.4"
 num-traits = "0.2"
